+++ src/main/java/com/auction/models/Bidder.java
@@ -1,6 +1,7 @@
 package com.auction.models;
 
 import com.auction.server.factory.UserRole;
+import java.math.BigDecimal;
 
 public class Bidder extends User {
     private static final long serialVersionUID = 1L;
@@ -11,7 +12,7 @@ public class Bidder extends User {
     }
 
     // Create a Bidder instance from an existed Bidder in DB
-    public Bidder(String id, String username, String password, double balance) {
+    public Bidder(String id, String username, String password, BigDecimal balance) {
         super(id, UserRole.BIDDER, username, password, balance);
     }
 }
