+++ src/main/java/com/auction/models/Electronics.java
@@ -1,19 +1,20 @@
 package com.auction.models;
 
 import com.auction.server.factory.ItemType;
+import java.math.BigDecimal;
 
 public class Electronics extends Item {
     private static final long serialVersionUID = 1L;
     private String brand;
 
     // Create a new Electronics instance
-    public Electronics(String itemName, String description, double startingPrice, String Brand, Seller owner){
+    public Electronics(String itemName, String description, BigDecimal startingPrice, String Brand, Seller owner){
         super(ItemType.ELECTRONICS, itemName,description,startingPrice, owner);
         this.brand = Brand;
     }
 
     // Create an Electronics instance from the database
-    public Electronics(String id, String itemName, String description, double startingPrice, String Brand, Seller owner){
+    public Electronics(String id, String itemName, String description, BigDecimal startingPrice, String Brand, Seller owner){
         super(id, ItemType.ELECTRONICS, itemName,description,startingPrice, owner);
         this.brand = Brand;
     }
